flask/requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
blinker==1.9.0
cffi==2.0.0
click==8.3.0
Flask==3.1.2
gitdb==4.0.12
GitPython==3.1.41
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
psycopg2-binary==2.9.10
pycparser==2.23
setuptools==75.6.0
smmap==5.0.2
Werkzeug==3.1.3
